Remove config-chart-bucket task
This task enables versioning and sets cache-control headers on helm
charts stored on GCS.

It's unclear what this adds. The bucket has versioning enabled and the
GCS resource is configured to version the relevant files. Cache-control
headers might come back to bite, but at that time it might be
easier/cleaner to open a PR on the GCS resource.

[Github #38]