Remove docker requirement for FATS
Previously the basis of FATS was running 'riff build'. In turn 'riff
build' spits out a temporary Dockerfile and sends this to a docker
daemon to be built. This works fine in a local dev environment, but
running dockerd inside of Garden containers is fraught with peril.

In c538d2f7a, responsibility for building the temporary Dockerfile
generated by a 'riff init' command is taken up by Concourse. This means
that FATS can instead run 'riff apply', if an image is available to pull
in.

This commit correlates with another commit in the main projectriff/riff
repo.

This commit does not represent the end-goal state: it deliberately uses
ginkgo's --focus parameter to only run the Node/Square test. This is
because only this has been set up elsewhere so far.

- Remove dockerd launch
- Install ginkgo
- Remove test-fail / reporting code, this can be done with fly
  intercept

[Github projectriff/ci#50]