Convert from erb to raw yaml
ERb (ie. ruby-based templates) are good at abstracting repetitive items,
but they make root-and-branch refactoring a bit hard to comprehend. This
commit simply checks in the "rendered" YAML generated by ./set_pipeline,
with some additional changes.

This commit is expected to be "broken" until the git repository settings
are reorganised to reflect the new monorepo introduced in
projectriff/riff#296.

- Convert to YAML
- Add header comments
- Convert existing header comments to uniform style
- Remove manual-publish-crd
- Remove manual-images-release
- Remove unnecessary 'master' branch config from git repos
- Remove git-kubernetes-model

[Github #52]